tests/examples: Fix the luacov import
Before this, this code always tried to load luacov, even when no luacovpath was
given. The result was that tests were run under luacov without applying our
configuration and thus more code ended up in the output than we intended.
Specifically, the tests code showed up on codecov as well.

Fix this by only loading luacov if we should do so.

A side effect of this is that things now properly fail if we should use luacov,
but it is not available.

Signed-off-by: Uli Schlachter <psychon@znc.in>