awesome-client: Handle errors when running arguments
Commit 78abb4a54cf58db21396cedc34f0aa46827212ff made awesome-client fail with a
non-zero exit code when sending a command that it got from stdin failed.

Commit f0f31bc305881aa0df54d66477843ffd475cb316 added the possibility to specify
commands to run as arguments to awesome-client. However, the exit code was still
zero even when such a command failed.

This commit makes awesome-client signal errors with its exit code even for code
specified via arguments.

Note that this means that following arguments will not be executed if some
argument fails. I do not know if this is the best behaviour or not, but I am
implementing it like this here due to its simplicity to implement.

Signed-off-by: Uli Schlachter <psychon@znc.in>