Fix short flicker when new clients appear
We are doing more and more things lazily in the C code. The newest
addition is lazily configuring clients, which means that geometry
changes are only applied later.

However, this caused a short flicker when a new client appears: We were
first making the client visible and then moving it to its "proper"
position.

Since unbanning is also done lazily, we just have to change the order in
which we apply these operations to fix this.

Signed-off-by: Uli Schlachter <psychon@znc.in>