[WIP] doc: Add a SVG screenshot utility.
Allows to take vectorial screenshots during tests. This, in turn,
make possible to dissect the screenshot using tools such as
Inscape or a Web browser introspection tool.

The purpose of this is being able to export some elements from
the integration tests into the documentation. It could have been
done using external tools, but how cool is that compared to SVG
screenshots.