_multi_screen: Help the garbage collector
Tests using _multi_screen end up with lots of screens, which also means
that they end up with lots of wiboxes, widgets and so on. Since
drawables are kept around in a weak table and are forced to redraw when
screens change, this causes slowdowns later on when we end up with lots
of drawables.

Help the garbage collector in cleaning up by giving it a slight kick in
the right moment.

Signed-off-by: Uli Schlachter <psychon@znc.in>