Remove menubar.get [API break]
The function seems useless and its documentation is wrong. It does not
return a wibox, but a widget. Also, the widget cannot really be used on
its own since it depends on the size of the wibox.

So menubar.get() and its wrapper through the metatables __call should
just be removed.

Signed-off-by: Uli Schlachter <psychon@znc.in>