Use only_on_screen in the declarative layout example
This is a follow-up to 052cda939b8a5ce1. Expressions like 's.index == 2'
and 's == screen.primary' might work at first. However, they are not
"dynamic": If screens are added/removed or the primary screen changes,
then the widgets are not updated to follow this.

Instead, the example now uses the new awful.widget.only_on_screen
container which provides the needed dynamic behaviour.

This is arguably "the real fix" to #1562 and #1565.

Signed-off-by: Uli Schlachter <psychon@znc.in>