Fix docs/07-my-first-awesome.md
In awesome 3.5, the default themes set a global variable called "theme".
This was, I think, not intentional and we changed this when luacheck
found this issue. However, the "my first awesome" document from the old
wiki made use of this fact and so many people were using this hack. This
commit fixes our copy of "my first awesome".

Signed-off-by: Uli Schlachter <psychon@znc.in>